UNIMOODP31-16:  models can not be edited when there are associated certificates emitted

+++ classes/external/deletemodel_external.php
@@ -30,7 +30,7 @@
  */
 
 namespace mod_certifygen\external;
-
+global $CFG;
 use external_api;
 use external_function_parameters;
 use external_single_structure;
@@ -38,7 +38,7 @@ use external_value;
 use invalid_parameter_exception;
 use mod_certifygen\persistents\certifygen_model;
 use moodle_exception;
-
+require_once($CFG->dirroot . '/mod/certifygen/lib.php');
 class deletemodel_external extends external_api {
     /**
      * Describes the external function parameters.
@@ -65,6 +65,10 @@ class deletemodel_external extends external_api {
         );
         $result = ['result' => true, 'message' => 'OK'];
         try {
+            $cemited = mod_certifygen_are_there_any_certificate_emited($id);
+            if ($cemited) {
+                return ['result' => false, 'message' => get_string('cannotdeletemodelcertemited', 'mod_certifygen')];
+            }
             $model = new certifygen_model($id);
             $model->delete();
         } catch (moodle_exception $e) {
